update structure project (modularization)
