+++ src/css/styles.css
@@ -6,3 +6,16 @@
 main {
   display: flex;
 }
+.btn-social-icon {
+  border-radius: 50%;
+  background: transparent;
+  width: 40px;
+  height: 40px;
+  border: 0px;
+  outline: none;
+}
+
+.btn-social-icon:hover {
+  box-shadow: 1px;
+  transform: translateY(-0.5px);
+}
