+++ src/index.html
@@ -21,7 +21,6 @@
 <footer></footer>
 <script src="https://www.gstatic.com/firebasejs/6.2.0/firebase-app.js"></script>
 <script  src="https://www.gstatic.com/firebasejs/6.2.0/firebase-auth.js"></script>
-<script type="module" src="index.js"></script>
 <script type="module" src="main.js"></script>
 </body>
 </html>
