--- src/model/login-model.js
@@ -1,49 +1,0 @@
-export const signInUserEmail = (email, password) => {
-  firebase
-    .auth()
-    .signInWithEmailAndPassword(email, password)
-    .then((result) => {
-      if (result.user.emailVerified) {
-        console.log(firebase.auth.currentUser.user);
-        console.log(firebase.auth().currentUser.uid);
-        window.location.hash = '#/home';
-      } else {
-        firebase.auth.signOut();
-        console.log('Por favor realiza la verificación en tu correo');
-      }
-    }).catch((error) => {
-      // Handle Errors here.
-      const errorCode = error.code;
-      const errorMessage = error.message;
-      console.log(errorCode);
-      console.log(errorMessage);
-    });
-};
-
-
-export const signInUserGoogle = () => {
-  const provider = new firebase.auth.GoogleAuthProvider();
-  firebase
-    .auth()
-    .signInWithPopup(provider)
-    .then((result) => {
-      window.location.hash = '#/home';
-      // This gives you a Google Access Token. You can use it to access the Google API.
-      const token = result.credential.accessToken;
-      console.log(token);
-      // The signed-in user info.
-      const user = result.user;
-      console.log(user);
-      // ...
-    }).catch((error) => {
-      // Handle Errors here.
-      const errorCode = error.code;
-      const errorMessage = error.message;
-      console.log(errorCode, errorMessage);
-      // The email of the user's account used.
-      const email = error.email;
-      // The firebase.auth.AuthCredential type that was used.
-      const credential = error.credential;
-      console.log(email, credential);
-    });
-};
