+++ src/routes/index.js
@@ -1,6 +1,6 @@
-import { changeView } from './routes/router.js';
+import { changeView } from './router.js';
 
-const init = () => {
+export const init = () => {
   changeView(window.location.hash);
   window.addEventListener('hashchange', () => (changeView(window.location.hash)));
 };
